add flatpaks to init

+++ os/butler.sh
@@ -53,8 +53,47 @@ set_theme() {
 
 # -- Init new install
 init() {
+  # Check if init already ran
+  if test -e "$HOME"/.config/apparatus/init-done; then
+    echo '{{ Bold "Already initialized.."}}' | gun format -t template
+    exit 0
+  fi
+  (
+  # Alacritty config
   mkdir -p ${HOME}/.config/alacritty
   cp /usr/share/apparatus/alacritty/* ~/.config/alacritty/
+
+  # FLATPAK - Install Firefox
+  echo "# Installing Firefox"
+  /usr/bin/flatpak install --user --noninteractive flathub org.mozilla.firefox
+  if [ "$?" != 0 ] ; then
+        zenity --error \
+          --text="Installing Firefox Failed"
+        exit 1
+  fi
+  echo "10"
+
+  echo "# Installing Signal"
+  /usr/bin/flatpak install --user --noninteractive flathub org.signal.Signal
+  if [ "$?" != 0 ] ; then
+        zenity --error \
+          --text="Installing Signal Failed"
+        exit 1
+  fi
+  echo "50"
+
+  echo "# Installing Joplin"
+  /usr/bin/flatpak install --user --noninteractive flathub net.cozic.joplin_desktop
+  if [ "$?" != 0 ] ; then
+        zenity --error \
+          --text="Installing Joplin Failed"
+        exit 1
+  fi
+  echo "100"
+
+
+  ) |
+    zenity --progress --title="uBlue Desktop Firstboot" --percentage=0 --auto-close --no-cancel --width=300
 }
 
 # Main
